feat(TransactionRow): Show txhash that links to block explorer
- also remove unused components
